Larry tidying up + parallelisation
I've made a few different files to separate things out and make it easier to read, untested as of yet so may need tweaking on the real machine. Also made  a file that attempts parallelisation of the control and the plotting of O2. Very crude currently and doesn't even allow for live control but I think that will be easy to implement from here